replace old Production implementations by new ones
