+++ src/arborator-modules/__models/pineTree.ts
