+++ src/arborator-modules/__productions/rules/GRule.ts
