+++ src/arborator-modules/__productions/rules/KRule.ts
