+++ src/arborator-modules/__productions/rules/TRule.ts
