+++ src/arborator-modules/productions/GRule.ts
@@ -1,0 +1,22 @@
+import Production from "../../lib/lsys/core/production";
+import { Glyph, Rule } from "../../lib/lsys/lsys";
+
+
+
+
+
+class GRule extends Production {
+
+	constructor( glyph: Rule ) {
+
+		super( glyph, [ glyph ] );
+	}
+
+	public process() {
+
+
+	}
+
+}
+
+export default GRule;
