+++ src/arborator-modules/productions/IRule.ts
@@ -1,0 +1,22 @@
+import Production from "../../lib/lsys/core/production";
+import { Glyph, Rule } from "../../lib/lsys/lsys";
+
+
+
+
+
+class IRule extends Production {
+
+
+	constructor( glyph: Rule, rule: Glyph[] ) {
+
+		super( glyph, rule );
+	}
+
+	process() {
+
+
+	}
+}
+
+export default IRule;
